Allow callback to shortcode parser to ignore unknown shortcodes.
With this patch, `Parser` takes an optional Boolean that tells it
whether to ignore unknown shortcodes or not. This is useful when
the parser is being used to collect information rather than translate
shortcodes, as it means the collecting parser doesn't need to have
handlers for all the shortcodes in the doc.